requirements.txt
requests==2.28.1
schedule==1.1.0
python-dotenv==0.21.0
pywin32>=306; sys_platform == 'win32'
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
APScheduler==3.9.1
Werkzeug==2.2.3 